feat(filter): add filtring by category

+++ database/migrations/2021_02_13_123647_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->float('price',);
+            $table->double('price',);
             $table->string('image');
             $table->timestamps();
         });
